+++ src/main/java/com/zerobase/moy/repository/jpa/UserRepository.java
@@ -1,4 +1,4 @@
-package com.zerobase.moy.repository;
+package com.zerobase.moy.repository.jpa;
 
 import com.zerobase.moy.data.entity.User;
 import java.util.Optional;
